+++ source/renderer/app/actions/hardware-wallets-actions.js
@@ -1,9 +1,5 @@
 // @flow
 import Action from './lib/Action';
-import type {
-  SetHardwareWalletLocalDataRequestType,
-  SetHardwareWalletDeviceRequestType,
-} from '../api/utils/localStorage';
 import type { DelegationAction } from '../types/stakingTypes';
 
 export default class HardwareWalletsActions {
@@ -20,8 +16,4 @@ export default class HardwareWalletsActions {
   }> = new Action();
   sendMoney: Action<any> = new Action();
   refreshHardwareWalletsLocalData: Action<any> = new Action();
-  setHardwareWalletLocalData: Action<SetHardwareWalletLocalDataRequestType> = new Action();
-  unsetHardwareWalletLocalData: Action<{ walletId: string }> = new Action();
-  setHardwareWalletDevice: Action<SetHardwareWalletDeviceRequestType> = new Action();
-  unsetHardwareWalletDevice: Action<{ deviceId: ?string }> = new Action();
 }