+++ source/renderer/app/stores/HardwareWalletsStore.js
@@ -174,18 +174,6 @@ export default class HardwareWalletsStore extends Store {
     hardwareWalletsActions.refreshHardwareWalletsLocalData.listen(
       this._refreshHardwareWalletsLocalData
     );
-    hardwareWalletsActions.setHardwareWalletLocalData.listen(
-      this._setHardwareWalletLocalData
-    );
-    hardwareWalletsActions.unsetHardwareWalletLocalData.listen(
-      this._unsetHardwareWalletLocalData
-    );
-    hardwareWalletsActions.setHardwareWalletDevice.listen(
-      this._setHardwareWalletDevice
-    );
-    hardwareWalletsActions.unsetHardwareWalletDevice.listen(
-      this._unsetHardwareWalletDevice
-    );
     getHardwareWalletConnectionChannel.onReceive(
       this._changeHardwareWalletConnectionStatus
     );
@@ -363,7 +351,7 @@ export default class HardwareWalletsStore extends Store {
 
   // @TODO - move to Transactions store once all logic fit and hardware wallets listed in general wallets list
   selectCoins = async (params: CoinSelectionsPaymentRequestType) => {
-    const { walletId, address, amount } = params;
+    const { walletId, address, amount, assets } = params;
     const wallet = this.stores.wallets.getWalletById(walletId);
     if (!wallet)
       throw new Error('Active wallet required before coins selections.');
@@ -374,6 +362,7 @@ export default class HardwareWalletsStore extends Store {
         payments: {
           address,
           amount,
+          assets,
         },
       });
       runInAction('HardwareWalletsStore:: set coin selections', () => {